AssertIsNotNone changed, and html site changed
setUp: Changed the html site to wikipedia, because there were problems
with the xpath on google.

test_checkContent: assertIsNotNone didn't work because [] is not None so
it would basically return true on everything, fixed by using assertTrue
and checking that the length is greater than 0.

test_lxmlNamespace: assertIsNotNone changed here too, but to
assertIsInstance, as I need to be sure it converts to the correct type.